moveit_backend/requirements.txt
django
djangorestframework
pillow
channels[daphne]
channels_redis
django-cors-headers